+++ kernel/Cargo.toml
@@ -8,6 +8,7 @@ edition.workspace = true
 bitflags = "2.11.0"
 riscv = { workspace = true }
 ksync = { workspace = true }
+heapless = "0.9.2"
 
 [build-dependencies]
 cc = "1"
