+++ kernel/src/task/mod.rs
@@ -1,0 +1,9 @@
+mod address_space;
+mod id;
+mod state;
+mod task;
+
+pub use address_space::*;
+pub use id::*;
+pub use state::*;
+pub use task::*;
