+++ kernel/src/task/state.rs
@@ -1,0 +1,10 @@
+#[repr(u8)]
+#[derive(Copy, Clone, PartialEq, Eq)]
+pub enum TaskState {
+    Sleeping,
+    Running,
+    Ready,
+    Blocked,
+    Zombie,
+    Exited,
+}
